Avoid deadlock in Close()
A call to listenerRoutine while closing can result in a deadlock. The
cause appears to be listenerRoutine not giving up when faced with an
error of type ERROR_OPERATION_ABORTED.

The change adds a case to specifically handle this error type and bubble
up ErrFileClosed to listenerRoutine, causing it to terminate gracefully.