Retry named pipe connection on race condition
If the named pipe server does not have an outstanding call to
`ConnectNamedPipe()` at the time the client calls `CreateFile()`
or `WaitNamedPipe()`, the client can receive `ERROR_FILE_NOT_FOUND`
and `ERROR_BAD_PATHNAME` errors.  When that's the case, there is
no choice but to retry after a short delay in the hope that next
time, the server will be able to accept the connection.

This error occurs infrequently in low traffic situations, but it
manifests quite reliably when clients open concurrent connections
and/or open many short-lived connections in a short period of
time.