Reuse pipe instances if listenerRoutine fails to create new instances
This commit enforces the invariant that a win32PipeListener guarantees
its stake in the pipe namespace by having instances of win32Pipe
"donate" their disconnected pipe instances to the listener if the
listener failed to create a new instance.

Currently, this is done with heavy use of sync.Mutex. The "best case"
of the listener successfully creating a successor instance elides one
locking call with the use of atomic.LoadPointer; it may be possible to
use atomic.CompareAndSwapPointer to achieve the same effect in general.
This is worth pursuing further.

This commit also fixes a race condition in win32PipeListener.Close,
where the close channel send may have been swallowed by
connectServerPipe if a client successfully connected before its pipe
instance was shut down.