Tests for Volume Mount Point API wrappers
These are a separate commit, because in order to implement the tests, a
bunch of new modules were pulled into go.mod.

And the tests only run as Administrator anyway.

On the plus side, the tests exercise a few APIs in the vhd package:
- CreateVirtualDisk
- AttachVirtualDisk
- DetachVirtualDisk
- GetVirtualDiskPhysicalPath

And in future, these tests could also exercise the DecodeReparsePoint
API too, since it should be 1:1 with GetVolumeNameForVolumeMountPoint.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>