pkg/guid: make code windows-only
When attempting to update go modules in a project, "go get" failed, because the
pkg/guid package contained Windows-only code, but was not marked as "windows-only",
resulting in the following error:

    GO111MODULE=on go get github.com/Microsoft/hcsshim@v0.8.14
    ....
    package github.com/Microsoft/hcsshim
        imports github.com/Microsoft/go-winio/pkg/guid
        imports golang.org/x/sys/windows: build constraints exclude all Go files in /go/pkg/mod/golang.org/x/sys@v0.0.0-20201201145000-ef89a241ccb3/windows

This patch marks the code as "windows-only", excluding it from other platforms.
A "doc.go" file was added so that the package is not considered "empty" on other
platforms.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>