Assign ErrPipeListenerClosed to net.ErrClosed
This error used to just be a `errors.New` with the text exactly matching
the "use of closed network connection" that you'd get from the Go stdlib.
Now that that error was exported as `net.ErrClosed` in Go 1.16 we should
be able to swap to this safely, and anyone who relied on checking the
string explicitly should still be fine.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>