go.mod: remove golang.org/x/tools dependency
This removes golang.org/x/tools/cmd/stringer as a direct dependency of this
module, instead using "go install" to install the required tool as part of
`go generate`. I opted for pinning to a fixed version for reproducibility, but
we could switch to `@latest` if there's no concerns for getting newer versions.

With this patch:

    which stringer
    # (no output)

    go generate ./pkg/guid
    which stringer
    /Users/thajeztah/go/bin/stringer

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>