tools/mkwinsyscall: replace deprecated funcs with SyscallN (drops go1.17)
The "syscall.Syscall<number>" functions were deprecated in go1.18 in favor
of the "syscall.SyscallN" function, which does not need the "nargs" argument,
and does not need the list of arguments to be padded with zeros.
https://github.com/golang/go/blob/go1.18/src/syscall/dll_windows.go#L27-L45

Now that go1.17 reached EOL and is no longer maintained, we can update the
code to use the new SyscallN function. This patch updates the mkwinsyscall
utility to generate code using the new SyscallN function, and removes the
utilities that are now redundant.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>